pyproject.toml
[project]
name = "llm-sheet-analyzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cachetools>=7.0.5",
    "fastapi[standard]>=0.135.3",
    "google-auth>=2.49.2",
    "gspread>=6.2.1",
    "openai>=2.31.0",
    "pandas>=3.0.2",
    "pytest>=9.0.3",
    "python-telegram-bot>=22.7",
    "structlog>=25.5.0",
]
